gemini/agents/genai-experience-concierge/pyproject.toml
[project]
name = "genai-experience-concierge"
version = "0.1.0"
description = "Example end-to-end agent deployment with session management and demo UI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-bigquery-storage>=2.30.0",
]

[project.scripts]
concierge = "scripts.cli.concierge:concierge"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.29.0",
    "pre-commit>=4.0.1",
    "mypy>=1.15.0",
    "ruff>=0.11.0",
    "types-requests>=2.32.0.20250306",
    "types-pyyaml>=6.0.12.20241230",
    "pip>=25.0.1",
    "pylint>=3.3.6",
    "google-adk>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./scripts"]

# leave mypy checks to demo sub-projects
[tool.mypy]
packages = ["scripts"]
disable_error_code = ["import-untyped"]

[tool.pylint.messages_control]
disable = ["wrong-import-order"]
